refactor(solver): split instantiation display helpers
